refactor: moved negative tests to its own file
